feat(play/local): Add subscription entitlements
Next thing I want to work on is user claims. Please refer to the nuxtconfig. In the session (user) info, there is a "subscriptions" field. I'd like to implement this as user claims. Admin can add one or more subscriptions, with the given statuses, and then they should show up in the token claims as well as the session respons